Inherit / merge gemini chat classes so all the state are shared and avoid some performance issues and regressions
