--- cogs/gemini/chat_mgmt.py
@@ -1,114 +1,0 @@
-from core.ai.core import ModelsList
-from core.ai.history import History
-from discord.ext import commands
-from os import environ
-import discord
-import motor.motor_asyncio
-
-class ChatMgmt(commands.Cog):
-    def __init__(self, bot):
-        self.bot = bot
-        self.author = environ.get("BOT_NAME", "Jakey Bot")
-
-        # Load the database and initialize the HistoryManagement class
-        # MongoDB database connection for chat history and possibly for other things
-        try:
-            self.HistoryManagement: History = History(db_conn=motor.motor_asyncio.AsyncIOMotorClient(environ.get("MONGO_DB_URL")))
-        except Exception as e:
-            raise e(f"Failed to connect to MongoDB: {e}...\n\nPlease set MONGO_DB_URL in dev.env")
-
-    ###############################################
-    # Clear context command
-    ###############################################
-    @commands.slash_command(
-        contexts={discord.InteractionContextType.guild, discord.InteractionContextType.bot_dm},
-        integration_types={discord.IntegrationType.guild_install, discord.IntegrationType.user_install}
-    )
-    async def sweep(self, ctx):
-        """Clear the context history of the conversation"""
-        await ctx.response.defer()
-
-        # Check if SHARED_CHAT_HISTORY is enabled
-        if environ.get("SHARED_CHAT_HISTORY", "false").lower() == "true":
-            guild_id = ctx.guild.id if ctx.guild else ctx.author.id
-        else:
-            guild_id = ctx.author.id
-
-        # This command is available in DMs
-        if ctx.guild is not None:
-            # This returns None if the bot is not installed or authorized in guilds
-            # https://docs.pycord.dev/en/stable/api/models.html#discord.AuthorizingIntegrationOwners
-            if ctx.interaction.authorizing_integration_owners.guild == None:
-                await ctx.respond("🚫 This commmand can only be used in DMs or authorized guilds!")
-                return  
-
-        # Initialize history
-        _feature = await self.HistoryManagement.get_config(guild_id=guild_id)
-
-        # Clear and set feature
-        await self.HistoryManagement.clear_history(guild_id=guild_id)
-        await self.HistoryManagement.set_config(guild_id=guild_id, tool=_feature)
-
-        await ctx.respond("✅ Chat history reset!")
-
-    # Handle errors
-    @sweep.error
-    async def on_application_command_error(self, ctx: discord.ApplicationContext, error: discord.DiscordException):
-        # Get original error
-        _error = getattr(error, "original")
-        if isinstance(_error, PermissionError):
-            await ctx.respond("⚠️ An error has occured while clearing chat history, logged the error to the owner")
-        elif isinstance(_error, FileNotFoundError):
-            await ctx.respond("ℹ️ Chat history is already cleared!")
-        else:
-            await ctx.respond("❌ Something went wrong, please check the console logs for details.")
-            raise error
-
-    ###############################################
-    # Set chat features command
-    ###############################################
-    @commands.slash_command(
-        contexts={discord.InteractionContextType.guild, discord.InteractionContextType.bot_dm},
-        integration_types={discord.IntegrationType.guild_install, discord.IntegrationType.user_install}
-    )
-    @discord.option(
-        "capability",
-        description = "Integrate tools to chat! Setting chat features will clear your history!",
-        choices=ModelsList.get_tools_list(),
-    )
-    async def feature(self, ctx, capability: str):
-        """Enhance your chat with capabilities! Some are in BETA so things may not always pick up"""
-        # Defer
-        await ctx.response.defer()
-
-        # Check if SHARED_CHAT_HISTORY is enabled
-        if environ.get("SHARED_CHAT_HISTORY", "false").lower() == "true":
-            guild_id = ctx.guild.id if ctx.guild else ctx.author.id
-        else:
-            guild_id = ctx.author.id
-
-        # This command is available in DMs
-        if ctx.guild is not None:
-            # This returns None if the bot is not installed or authorized in guilds
-            # https://docs.pycord.dev/en/stable/api/models.html#discord.AuthorizingIntegrationOwners
-            if ctx.interaction.authorizing_integration_owners.guild == None:
-                await ctx.respond("🚫 This commmand can only be used in DMs or authorized guilds!")
-                return
-
-        # if tool use is the same, do not clear history
-        _feature = await self.HistoryManagement.get_config(guild_id=guild_id)
-        if _feature == capability:
-            await ctx.respond("✅ Feature already enabled!")
-        else:
-            # set config
-            await self.HistoryManagement.set_config(guild_id=guild_id, tool=capability)
-            await ctx.respond(f"✅ Feature **{capability}** enabled successfully and chat is reset to reflect the changes")
-        
-    # Handle errors
-    @feature.error
-    async def on_application_command_error(self, ctx: discord.ApplicationContext, error: discord.DiscordException):
-        await ctx.respond("❌ Something went wrong, please check the console logs for details.")
-        raise error
-
-def setup(bot):
-    bot.add_cog(ChatMgmt(bot))
